requirements.txt
Flask
whisper
gtts
google-generativeai
werkzeug
pydub
numpy
